// src/types/product.ts
export interface Product {
    id: number;
    name: string;
    image: string;
    price: number;
    categoryId: string[];
    description?: string;
    sale?: Sale;
    variants?: Variant[];
}

/**
 * Interface Sale
 */
export interface PercentSale {
    type: "percent";
    percent: number;
}

export interface FixedSale {
    amount: number;
    type: "fixed";
}

export type Sale = PercentSale | FixedSale;

/**
 * Interface Variant
 */
export interface Option {
    id: string;
    label?: string;
    priceChange?: Sale;
}

export interface BaseVariant {
    id: string;
    label?: string;
    options: Option[];
}

export interface SingleOptionVariant extends BaseVariant {
    type: "single";
    default?: string;
}

export interface MultipleOptionVariant extends BaseVariant {
    type: "multiple";
    default?: string[];
}

export type Variant = SingleOptionVariant | MultipleOptionVariant;